Migrate Value::Set methods in chrome/browser/extensions/api/
The Value::SetKey and Value::SetPath families of methods are depreacated
in favor of the Value::Dict::Set() and Value::Dict::SetByDottedPath()
methods, respectively.

The only SetBoolPath() call in this directory doesn't actually use a key
with any dots in it, so migrate it to Set() instead of SetByDottedPath()

Bug: 1187001
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4289444
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1110644}

+++ chrome/browser/extensions/api/page_capture/page_capture_api.h
@@ -59,7 +59,7 @@ class PageCaptureSaveAsMHTMLFunction : public ExtensionFunction {
 
   // Called on the UI thread.
   void ReturnFailure(const std::string& error);
-  void ReturnSuccess(int64_t file_size);
+  void ReturnSuccess(int file_size);
 
   // Callback called once the MHTML generation is done.
   void MHTMLGenerated(int64_t mhtml_file_size);
